==> scripts/Step05_cite_packages.R <==
## ---------------------------
## Author: Christine A. Parisek, caparisek@ucdavis.edu
## This script: Gets citations for the unique packages used in this project 
## ---------------------------

?citation
?packageVersion

packageVersion("tidyverse")   # v2.0.0
packageVersion("data.table")  # v1.14.8
packageVersion("tigris")      # v2.0.3
packageVersion("sf")          # v1.0.12
packageVersion("sp")          # v1.6.0     
packageVersion("ggspatial")   # v1.1.8
packageVersion("mgcv")        # v1.8.42
packageVersion("MuMIn")       # v1.47.5
packageVersion("lmerTest")    # v3.1.3
packageVersion("smatr")       # v3.4.8
packageVersion("fBasics")     # v4022.94
packageVersion("patchwork")   # v1.1.2
packageVersion("cowplot")     # v1.1.1
packageVersion("LaCroixColoR")# v0.1.0
packageVersion("scales")      # v1.2.1

citation("tidyverse")
citation("sf")
citation("sp")         
citation("ggspatial")
citation("tigris")
citation("mgcv")
citation("MuMIn")
citation("lmerTest")
citation("smatr")
citation("fBasics")
citation("data.table")
citation("scales")
citation("patchwork")
citation("cowplot")
citation("LaCroixColoR")




# {tidyverse} v2.0.0
# {data.table} v1.14.8
# {tigris} v2.0.3
# {sf} v1.0.12
# {sp} v1.6.0     
# {ggspatial} v1.1.8
# {mgcv} v1.8.42
# {MuMIn} v1.47.5
# {lmerTest} v3.1.3
# {smatr} v3.4.8
# {fBasics} v4022.94
# {patchwork} v1.1.2
# {cowplot} v1.1.1
# {LaCroixColoR} v0.1.0
# {scales} v1.2.1
